TextDocument into own file; TextEditor into views/
